Started removing the God Class <feature
